Fix all 45 build warnings (now 0 warnings)
- C4267 size_t→int: add explicit (int) casts for .size(), strlen(), wcslen()
- C4244 LONG_PTR→LONG: change lastWindowStyle/Ex to LONG_PTR
- C4244 LRESULT→int/DWORD: add explicit casts for SendMessage returns
- C4244 WPARAM→int: add explicit casts for character arithmetic
- C4477 format string: cast WPARAM to int for %c format specifier
- C4312 DWORD→void*: cast through uintptr_t for _beginthreadex param
- C4005 SelectFont: #undef before redefining to avoid wingdi.h conflict
- C4101/C4244 in ns-eel2: initialize delim, cast WDL_INT64 modulo to EEL_F

+++ src/audio/log.cpp
@@ -15,7 +15,7 @@ void LOGA(LPCSTR format, ...) {
   char output_buff[LOG_SIZE];
 
   char err[20] = ""; //"Warn: ";
-  int sl = strnlen_s(err, 20);
+  size_t sl = strnlen_s(err, 20);
   strncpy_s(output_buff, err, sl);
 
   va_list args_list;
@@ -31,7 +31,7 @@ void LOG(LPCWSTR format, ...) {
   wchar_t output_buff[LOG_SIZE];
 
   wchar_t err[20] = L""; //L"Warn: ";
-  int sl = wcsnlen_s(err, 20);
+  size_t sl = wcsnlen_s(err, 20);
   wcsncpy_s(output_buff, err, sl);
 
   va_list args_list;
@@ -49,7 +49,7 @@ void ERRA(LPCSTR format, ...) {
   char output_buff[LOG_SIZE];
 
   char err[20] = "Error: ";
-  int sl = strnlen_s(err, 20);
+  size_t sl = strnlen_s(err, 20);
   strncpy_s(output_buff, err, sl);
 
   va_list args_list;
@@ -65,7 +65,7 @@ void ERR(LPCWSTR format, ...) {
   wchar_t output_buff[LOG_SIZE];
 
   wchar_t err[20] = L"Error: ";
-  int sl = wcsnlen_s(err, 20);
+  size_t sl = wcsnlen_s(err, 20);
   wcsncpy_s(output_buff, err, sl);
 
   va_list args_list;
